Add support for encoding '+-infinity'
Erlang can't process the float that is equivialant of Javascripts Infinity
(Number.(POSITIVE,NEGATIVE)_INFINITY) which essentially is the max- and
minimum floating point.

Since there is no concept of a minimum value both are atoms.  This means
you can compare any number to '+infinity" and it will always be larger;
however '-infinity' will also be larger than any number.